[MOV] im_livechat: move CCThrottleFunctionObject outside of -
-public_livechat.js

Task-2212347
